Ticket #2752: CMake: sdlExtern
* fix 64-bit gcc build of wxIFM

git-svn-id: https://isrsvn.usurf.usu.edu/svn/isr/roots/wxWidgets/wxx/wxIFM@19290 a06e8daa-702c-0410-a579-eb5e6c4fa09d

+++ src/ifm/plugin.cpp
@@ -289,12 +289,12 @@ wxIFMComponent *wxIFMInterfacePluginBase::GetFloatingComponentByPos(const wxPoin
 void wxIFMInterfacePluginBase::AddChildWindow(wxWindow *window, wxIFMComponent *component)
 {
     // store the child window / component pair
-    m_windows[(int)window] = component;
+    m_windows[(size_t)window] = component;
 }
 
 wxIFMComponent *wxIFMInterfacePluginBase::FindChildWindow(wxWindow *window)
 {
-    return m_windows[(int)window];
+    return m_windows[(size_t)window];
 }
 
 void wxIFMInterfacePluginBase::OnMouseEvent(wxMouseEvent &event)
